Rename rule complexity command from analyze to complexity
