+++ internal/cmd/complexity/complexity.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by an Apache2
 // license that can be found in the LICENSE file.
 
-package analyze
+package complexity
 
 import (
 	"fmt"
